requirements.txt
alem==0.0.3.dev0
alembic==1.9.2
anyio==3.6.2
astroid==2.14.1
attrs==22.2.0
bcrypt==4.0.1
black==23.1.0
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
cryptography==39.0.0
dill==0.3.6
ecdsa==0.18.0
exceptiongroup==1.1.0
fastapi==0.89.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
iniconfig==2.0.0
isort==5.12.0
lazy-object-proxy==1.9.0
Mako==1.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.0
passlib==1.7.4
pathspec==0.11.0
pika==1.3.1
platformdirs==3.0.0
pluggy==1.0.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.4
pylint==2.16.1
PyMySQL==1.0.2
pytest==7.2.1
pytest-asyncio==0.20.3
python-dotenv==0.21.1
python-jose==3.3.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.1
starlette==0.22.0
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
uvicorn==0.20.0
wrapt==1.14.1
zipp==3.12.1
